python-code/project_files/requirements.txt
aiohttp
aiolimiter
asyncio
async-openai
fer
librosa
matplotlib
numpy
openai
opencv-python
pydub
python-dotenv
scipy
seaborn
tiktoken
tenacity
torch
transformers
